+++ libweatherapi/src/main/java/com/hartcode/hartweather/libweatherapi/Weather.java
@@ -15,8 +15,8 @@ public class Weather {
     public String description;
     public String icon;
     public float temp;
-    public int pressure;
-    public int humidity;
+    public float pressure;
+    public float humidity;
     public float temp_min;
     public float temp_max;
     public long lastUpdate;
@@ -26,7 +26,7 @@ public class Weather {
 
     }
 
-    public Weather(long id, int cityId, float lat, float lon, String cityName, String main, String description, String icon, float temp, int pressure, int humidity, float temp_min, float temp_max, long lastUpdate)
+    public Weather(long id, int cityId, float lat, float lon, String cityName, String main, String description, String icon, float temp, float pressure, float humidity, float temp_min, float temp_max, long lastUpdate)
     {
         this.id = id;
         this.cityId = cityId;
